Added Method that return PlaceModels so that the a Marking object can be referenced better. Fixed a bug in the CircleOrder when more than 360 objects where created
M. Fladt

+++ WoPeD-QualAnalysis/src/org/woped/qualanalysis/reachabilitygraph/data/CircleCoordinates.java
@@ -11,7 +11,7 @@ public class CircleCoordinates {
 		int mheight = (height / 2) - 45;
 
 		// get the steps in degrees
-		int degreestep = 360 / objectcount;
+		double degreestep = 360 / objectcount;
 		
 		//Fill the linked list with coordinates for every object
 		for(int i = 0; i < objectcount; i++) {
